Added hypridle and configs for lock after suspend and idle suspend and lock

+++ .config/wofi/powerMenu/powerMenu.sh
@@ -3,15 +3,21 @@
 # Define la ruta completa al comando systemctl
 SYSTEMCTL=/usr/bin/systemctl
 
+# Comprueba que se pase un argumento
+if [ -z "$1" ]; then
+    exit 1
+fi
+
 case "$1" in
     Block)
-        # Aquí puedes añadir el comando para bloquear la pantalla si lo deseas.
+        hyprlock
         ;;
     Shutdown)
         $SYSTEMCTL poweroff
         ;;
     "Log Out")
-        # Aquí puedes añadir el comando para cerrar sesión.
+        # Puedes cambiar esto por el comando para cerrar sesión que uses
+        $SYSTEMCTL reboot  # Cambiar a logout si es necesario
         ;;
     Suspend)
         $SYSTEMCTL suspend
